+++ src/main/java/com/product_comparator/productcomparator/entity/PriceHistory.java
@@ -1,4 +1,4 @@
-package com.product_comparator.productcomparator.model;
+package com.product_comparator.productcomparator.entity;
