+++ src/main/java/com/product_comparator/productcomparator/repository/DiscountRepository.java
@@ -1,0 +1,9 @@
+package com.product_comparator.productcomparator.repository;
+
+import com.product_comparator.productcomparator.entity.Discount;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface DiscountRepository extends JpaRepository<Discount,Integer> {
+}
