+++ src/main/java/com/product_comparator/productcomparator/service/DiscountService.java
@@ -1,0 +1,9 @@
+package com.product_comparator.productcomparator.service;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
+
+@Service
+public class DiscountService {
+    //@Autowired
+}
