+++ src/main/resources/application.properties
@@ -1,3 +1,12 @@
 spring.application.name=product-comparator
 csv.prices.path=classpath:data/*.csv
 csv.discounts.path=classpath:data/discounts/*.csv
+spring.datasource.url= jdbc:postgresql://localhost:5432/postgres
+spring.datasource.username= postgres
+spring.datasource.password= masface123
+
+spring.jpa.properties.hibernate.jdbc.lob.non_contextual_creation= true
+spring.jpa.properties.hibernate.dialect= org.hibernate.dialect.PostgreSQLDialect
+
+# Hibernate ddl auto (create, create-drop, validate, update)
+spring.jpa.hibernate.ddl-auto= update
